requirements.txt
Flask==2.3.2
requests==2.31.0
pandas==2.3.2
google-generativeai==0.4.1
python-dotenv==1.0.0
gunicorn==21.2.0